requirements.txt
sounddevice>=0.5.0
numpy<2.0
onnxruntime-directml
sentencepiece
keyboard
openai
huggingface_hub
pystray
Pillow
